+++ pkg/template/worker_template.go
@@ -182,6 +182,7 @@ systemd:
       Description=k8s-kubelet
       StartLimitIntervalSec=0
       [Service]
+      User=root
       TimeoutStartSec=300
       Restart=always
       RestartSec=0
@@ -364,6 +365,12 @@ storage:
       mode: 0644
       contents:
         source: "data:text/plain;charset=utf-8;base64,{{  index .Files "conf/docker-daemon.json" }}"
+
+    - path: /root/.docker/config.json
+      filesystem: root
+      mode: 0644
+      contents:
+        source: "data:text/plain;charset=utf-8;base64,{{  index .Files "conf/kubelet-docker-config.json" }}"
 
     - path: /etc/modules-load.d/ip_vs.conf
       filesystem: root