+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Added
+
+- Add dockerhub authentication for kubelet.
+- Use explicitly user `root` for kubelet systemd unit.
+
 ## [8.0.4] - 2020-10-21
 
 ### Changed